Move more logic to helpers from users and doctors controller
