+++ app/controllers/api/v1/users_controller.rb
@@ -1,4 +1,7 @@
 class Api::V1::UsersController < ApplicationController
+  include ApplicationHelper
+  include UserHelper
+
   before_action :set_user, only: %i[show update destroy]
   before_action :reject_attempt_to_create_doctor, only: %i[create]
 
@@ -39,12 +42,5 @@ class Api::V1::UsersController < ApplicationController
 
   def set_user
     @patient = User.patient.find(set_user_params[:id])
-  end
-
-  def reject_attempt_to_create_doctor
-    return unless params[:controller] == 'api/v1/users' && (params[:is_doctor] || params[:office_address])
-
-    render json: { error: "Can't create doctor from a /#{params[:controller]} endpoint" },
-           status: :forbidden
   end
 end
